+++ outputs.tf
@@ -30,6 +30,10 @@ output "ssl_alert_policy_names" {
   value = { for days, alert in google_monitoring_alert_policy.ssl_expiring_days : days => alert.name }
 }
 
+output "gke_node_count_alert_policy_name" {
+  value = one(google_monitoring_alert_policy.gke_node_count[*].name)
+}
+
 output "typesense_logmatch_alert_policy_names" {
   value = { for k, v in google_monitoring_alert_policy.typesense_logmatch_alert : k => v.name }
 }